refactor(exceptions): move truncate_text utility to _utils.py
